New Branch, attempting to merge
Cleaned up some file structure (two of our folders were named "Assets"
so the sprites folder was moved out of the secondary assets folder and
moved into the main assets folder) and added my scripts for the
Inventory system, pre-battle system, as well as the base for the Game
State system and the player/commander object. These are from the
Inventory-PreBattle branch I orginally created, but was unable to merge
due to conflicts.

This new branch will also be a test to see if the master branch has
become damaged.